Fix the breadcrumb problem, and display page titles for the new organization pages
