+++ src/components/MAPI/organizations/OrganizationIndex.tsx
@@ -1,11 +1,10 @@
-import Cluster from 'Cluster/Cluster';
 import * as React from 'react';
 import { Breadcrumb } from 'react-breadcrumbs';
 import { Redirect, Route, Switch } from 'react-router-dom';
 import { OrganizationsRoutes } from 'shared/constants/routes';
 import LoadingOverlay from 'UI/Display/Loading/LoadingOverlay';
 
-import OrganizationDetail from './OrganizationDetail';
+import Organization from './Organization';
 import OrganizationList from './OrganizationList';
 
 const OrganizationIndex: React.FC = () => {
@@ -20,11 +19,7 @@ const OrganizationIndex: React.FC = () => {
             exact
             path={OrganizationsRoutes.List}
           />
-          <Route component={Cluster} path={OrganizationsRoutes.Clusters.Home} />
-          <Route
-            component={OrganizationDetail}
-            path={OrganizationsRoutes.Detail}
-          />
+          <Route component={Organization} path={OrganizationsRoutes.Detail} />
           <Redirect to={OrganizationsRoutes.Home} />
         </Switch>
       </LoadingOverlay>
