+++ src/components/MAPI/organizations/OrganizationList.tsx
@@ -6,6 +6,7 @@ import { createOrganization } from 'model/services/mapi/securityv1alpha1/createO
 import * as React from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { OrganizationsRoutes } from 'shared/constants/routes';
+import DocumentTitle from 'shared/DocumentTitle';
 import { getLoggedInUser } from 'stores/main/selectors';
 import { selectOrganizations } from 'stores/organization/selectors';
 import OrganizationListPage from 'UI/Display/Organizations/OrganizationListPage';
@@ -18,43 +19,48 @@ const OrganizationIndex: React.FC = () => {
   const user = useSelector(getLoggedInUser);
 
   return (
-    <OrganizationListPage
-      onClickRow={(name) => {
-        const orgPath = RoutePath.createUsablePath(OrganizationsRoutes.Detail, {
-          orgId: name,
-        });
+    <DocumentTitle title='Organizations'>
+      <OrganizationListPage
+        onClickRow={(name) => {
+          const orgPath = RoutePath.createUsablePath(
+            OrganizationsRoutes.Detail,
+            {
+              orgId: name,
+            }
+          );
 
-        dispatch(push(orgPath));
-      }}
-      data={Object.keys(organizations).map((orgName) => ({
-        name: orgName,
-      }))}
-      // TODO: @oponder: Do we like this? Handling errors and doing requests and mutation in the component?
-      //                 I was generally always ok with it.. but it feels like I am breaking some rules.
-      createOrg={async (orgName) => {
-        if (user) {
-          try {
-            await createOrganization(client, user, orgName)();
-          } catch (error) {
-            if (error?.config?.data?.message) {
-              new FlashMessage(
-                `Unable to create organization "${orgName}"`,
-                messageType.ERROR,
-                messageTTL.LONG,
-                error.config.data.message
-              );
-            } else {
-              new FlashMessage(
-                `Unable to create organization "${orgName}"`,
-                messageType.ERROR,
-                messageTTL.LONG,
-                'Something unexpected went wrong while trying to create this organization'
-              );
+          dispatch(push(orgPath));
+        }}
+        data={Object.keys(organizations).map((orgName) => ({
+          name: orgName,
+        }))}
+        // TODO: @oponder: Do we like this? Handling errors and doing requests and mutation in the component?
+        //                 I was generally always ok with it.. but it feels like I am breaking some rules.
+        createOrg={async (orgName) => {
+          if (user) {
+            try {
+              await createOrganization(client, user, orgName)();
+            } catch (error) {
+              if (error?.config?.data?.message) {
+                new FlashMessage(
+                  `Unable to create organization "${orgName}"`,
+                  messageType.ERROR,
+                  messageTTL.LONG,
+                  error.config.data.message
+                );
+              } else {
+                new FlashMessage(
+                  `Unable to create organization "${orgName}"`,
+                  messageType.ERROR,
+                  messageTTL.LONG,
+                  'Something unexpected went wrong while trying to create this organization'
+                );
+              }
             }
           }
-        }
-      }}
-    />
+        }}
+      />
+    </DocumentTitle>
   );
 };
